Fix: rename utils modules
